+++ app/routes/web.php
@@ -2,13 +2,14 @@
 
 declare(strict_types=1);
 
+use App\Controllers\HomeController;
 use App\Controllers\UserController;
 use App\Middlewares\Csrf;
 use Monad\Clarity\Services\Request;
 use Monad\Clarity\Services\Route;
 use Monad\Clarity\Services\View;
 
-Route::get('/', fn () => View::render('Home/index'));
+Route::get('/', [HomeController::class, 'index']);
 
 Route::get('/users', [UserController::class, 'index']);
 Route::get('/users/create', [UserController::class, 'create']);